Fix isnan issues on Win

+++ plugins/ros1_parsers/ros1_parser.cpp
@@ -64,7 +64,7 @@ bool IntrospectionParser::parseMessage(MessageRef serialized_msg, double timesta
 
     auto& series = getSeries(key);
 
-    if (!std::isnan(value) && !std::isinf(value))
+    if (!std::isnan(static_cast<double>(value)) && !std::isinf(value))
     {
       series.pushBack({ timestamp, value });
     }
